moved auto out of commands and moved some other stuff
