+++ src/main/java/frc/robot/RobotContainer.java
@@ -4,9 +4,9 @@
 
 package frc.robot;
 
+import frc.robot.auto.AutoSelector;
 import frc.robot.commands.*;
 import frc.robot.subsystems.*;
-import frc.robot.commands.auto.AutoSelector;
 import edu.wpi.first.cameraserver.CameraServer;
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.button.CommandXboxController;
