+++ src/main/java/frc/robot/auto/AutonomousBalanceMobile.java
@@ -2,20 +2,21 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands.auto.AutoRoutines;
+package frc.robot.auto;
 
 import edu.wpi.first.wpilibj2.command.InstantCommand;
 import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
 import edu.wpi.first.wpilibj2.command.WaitCommand;
 import frc.robot.Constants;
-import frc.robot.commands.auto.AutoFunctions.*;
+import frc.robot.commands.Balance;
+import frc.robot.commands.DriveStraight;
 import frc.robot.subsystems.Bucket;
 import frc.robot.subsystems.DriveBase;
-import frc.robot.subsystems.Intake;
 import frc.robot.subsystems.Gyro;
+import frc.robot.subsystems.Intake;
 import edu.wpi.first.wpilibj.DoubleSolenoid;
 
-public class AutonomousBalanceNoMobile extends SequentialCommandGroup {
+public class AutonomousBalanceMobile extends SequentialCommandGroup {
   /*
    * pseudoCode:
    * humans will position robot
@@ -23,11 +24,12 @@ public class AutonomousBalanceNoMobile extends SequentialCommandGroup {
    * robot drives forward, getting more auto points
    */
 
-  // Subsystem to Dump Cargo then go forward over charge station
-  // and then back up onto charge system to attempt balance
-  public AutonomousBalanceNoMobile(DriveBase drive, Intake intake, Bucket bucket, Gyro gyro) {
+   // Subsystem to Dump Cargo then go forward over charge station
+   // and then back up onto charge system to attempt balance
+
+  public AutonomousBalanceMobile(DriveBase drive, Intake intake, Bucket bucket, Gyro gyro) {
     super(
-      //new WaitCommand(2),
+      new WaitCommand(2),
       //dump game piece
       new InstantCommand(
         ()->bucket.set(DoubleSolenoid.Value.kForward),
@@ -40,9 +42,12 @@ public class AutonomousBalanceNoMobile extends SequentialCommandGroup {
         bucket
       ),
       new WaitCommand(1),
-      //go forward onto charge station
-      new DriveStraight(drive, 2,Constants.auto.fwdSpeed),
-      // begin balancing
+      //go forward
+      new DriveStraight(drive, 2.7,Constants.auto.fwdSpeed),
+      new WaitCommand(1),
+      //go back
+      new DriveStraight(drive, 3,Constants.auto.revSpeed),
+      // balance
       new Balance(drive, gyro)
     );
   }
