+++ src/main/java/frc/robot/auto/AutonomousGrab.java
@@ -2,17 +2,19 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands.auto.AutoRoutines;
+package frc.robot.auto;
 
 import edu.wpi.first.wpilibj.DoubleSolenoid;
 import edu.wpi.first.wpilibj2.command.InstantCommand;
 import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
 import edu.wpi.first.wpilibj2.command.WaitCommand;
-import frc.robot.commands.auto.AutoFunctions.*;
 import frc.robot.Constants;
+
+import frc.robot.commands.DriveStraight;
 import frc.robot.subsystems.Bucket;
 import frc.robot.subsystems.DriveBase;
 import frc.robot.subsystems.Intake;
+
 
 public class AutonomousGrab extends SequentialCommandGroup {
