+++ src/main/java/frc/robot/commands/Balance.java
@@ -3,7 +3,7 @@
 // the WPILib BSD license file in the root directory of this project.
 //
 
-package frc.robot.commands.auto.AutoFunctions;
+package frc.robot.commands;
 
 import edu.wpi.first.math.MathUtil;
 import edu.wpi.first.math.controller.PIDController;
