+++ src/main/java/frc/robot/commands/DriveStraight.java
@@ -1,4 +1,4 @@
-package frc.robot.commands.auto.AutoFunctions;
+package frc.robot.commands;
 
 import edu.wpi.first.math.MathUtil;
 import edu.wpi.first.wpilibj.Encoder;
